+++ lib/input/constructor.go
@@ -89,6 +89,7 @@ const (
 	TypeSTDIN           = "stdin"
 	TypeTCP             = "tcp"
 	TypeTCPServer       = "tcp_server"
+	TypeUDPServer       = "udp_server"
 	TypeWebsocket       = "websocket"
 	TypeZMQ4            = "zmq4"
 )
@@ -127,6 +128,7 @@ type Config struct {
 	STDIN           STDINConfig                  `json:"stdin" yaml:"stdin"`
 	TCP             TCPConfig                    `json:"tcp" yaml:"tcp"`
 	TCPServer       TCPServerConfig              `json:"tcp_server" yaml:"tcp_server"`
+	UDPServer       UDPServerConfig              `json:"udp_server" yaml:"udp_server"`
 	Websocket       reader.WebsocketConfig       `json:"websocket" yaml:"websocket"`
 	ZMQ4            *reader.ZMQ4Config           `json:"zmq4,omitempty" yaml:"zmq4,omitempty"`
 	Processors      []processor.Config           `json:"processors" yaml:"processors"`
@@ -165,6 +167,7 @@ func NewConfig() Config {
 		STDIN:           NewSTDINConfig(),
 		TCP:             NewTCPConfig(),
 		TCPServer:       NewTCPServerConfig(),
+		UDPServer:       NewUDPServerConfig(),
 		Websocket:       reader.NewWebsocketConfig(),
 		ZMQ4:            reader.NewZMQ4Config(),
 		Processors:      []processor.Config{},