+++ lib/input/udp_server_test.go
@@ -35,22 +35,15 @@ import (
 	"github.com/Jeffail/benthos/lib/types"
 )
 
-func TestTCPBasic(t *testing.T) {
-	ln, err := net.Listen("tcp", "127.0.0.1:0")
-	if err != nil {
-		if ln, err = net.Listen("tcp6", "[::1]:0"); err != nil {
-			t.Fatalf("failed to listen on a port: %v", err)
-		}
-	}
-	defer ln.Close()
-
+func TestUDPServerBasic(t *testing.T) {
 	conf := NewConfig()
-	conf.TCP.Address = ln.Addr().String()
+	conf.UDPServer.Address = "127.0.0.1:0"
 
-	rdr, err := NewTCP(conf, nil, log.Noop(), metrics.Noop())
+	rdr, err := NewUDPServer(conf, nil, log.Noop(), metrics.Noop())
 	if err != nil {
 		t.Fatal(err)
 	}
+	addr := rdr.(*UDPServer).Addr()
 
 	defer func() {
 		rdr.CloseAsync()
@@ -59,7 +52,7 @@ func TestTCPBasic(t *testing.T) {
 		}
 	}()
 
-	conn, err := ln.Accept()
+	conn, err := net.Dial("udp", addr.String())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -81,10 +74,9 @@ func TestTCPBasic(t *testing.T) {
 	}()
 
 	readNextMsg := func() (types.Message, error) {
-		var msg types.Message
+		var tran types.Transaction
 		select {
-		case tran := <-rdr.TransactionChan():
-			msg = tran.Payload.DeepCopy()
+		case tran = <-rdr.TransactionChan():
 			select {
 			case tran.ResponseChan <- response.NewAck():
 			case <-time.After(time.Second):
@@ -93,7 +85,7 @@ func TestTCPBasic(t *testing.T) {
 		case <-time.After(time.Second):
 			return nil, errors.New("timed out")
 		}
-		return msg, nil
+		return tran.Payload, nil
 	}
 
 	exp := [][]byte{[]byte("foo")}
@@ -125,22 +117,15 @@ func TestTCPBasic(t *testing.T) {
 	conn.Close()
 }
 
-func TestTCPReconnect(t *testing.T) {
-	ln, err := net.Listen("tcp", "127.0.0.1:0")
-	if err != nil {
-		if ln, err = net.Listen("tcp6", "[::1]:0"); err != nil {
-			t.Fatalf("failed to listen on a port: %v", err)
-		}
-	}
-	defer ln.Close()
-
+func TestUDPServerReconnect(t *testing.T) {
 	conf := NewConfig()
-	conf.TCP.Address = ln.Addr().String()
+	conf.UDPServer.Address = "127.0.0.1:0"
 
-	rdr, err := NewTCP(conf, nil, log.Noop(), metrics.Noop())
+	rdr, err := NewUDPServer(conf, nil, log.Noop(), metrics.Noop())
 	if err != nil {
 		t.Fatal(err)
 	}
+	addr := rdr.(*UDPServer).Addr()
 
 	defer func() {
 		rdr.CloseAsync()
@@ -149,7 +134,7 @@ func TestTCPReconnect(t *testing.T) {
 		}
 	}()
 
-	conn, err := ln.Accept()
+	conn, err := net.Dial("udp", addr.String())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -163,24 +148,23 @@ func TestTCPReconnect(t *testing.T) {
 			t.Error(cerr)
 		}
 		conn.Close()
-		conn, cerr = ln.Accept()
+		conn, cerr = net.Dial("udp", addr.String())
 		if cerr != nil {
 			t.Fatal(cerr)
 		}
-		if _, cerr := conn.Write([]byte("bar\n")); cerr != nil {
+		if _, cerr = conn.Write([]byte("bar\n")); cerr != nil {
 			t.Error(cerr)
 		}
-		if _, cerr := conn.Write([]byte("baz\n")); cerr != nil {
+		if _, cerr = conn.Write([]byte("baz\n")); cerr != nil {
 			t.Error(cerr)
 		}
 		wg.Done()
 	}()
 
 	readNextMsg := func() (types.Message, error) {
-		var msg types.Message
+		var tran types.Transaction
 		select {
-		case tran := <-rdr.TransactionChan():
-			msg = tran.Payload.DeepCopy()
+		case tran = <-rdr.TransactionChan():
 			select {
 			case tran.ResponseChan <- response.NewAck():
 			case <-time.After(time.Second):
@@ -189,7 +173,7 @@ func TestTCPReconnect(t *testing.T) {
 		case <-time.After(time.Second):
 			return nil, errors.New("timed out")
 		}
-		return msg, nil
+		return tran.Payload, nil
 	}
 
 	exp := [][]byte{[]byte("foo")}
@@ -221,23 +205,16 @@ func TestTCPReconnect(t *testing.T) {
 	conn.Close()
 }
 
-func TestTCPMultipart(t *testing.T) {
-	ln, err := net.Listen("tcp", "127.0.0.1:0")
-	if err != nil {
-		if ln, err = net.Listen("tcp6", "[::1]:0"); err != nil {
-			t.Fatalf("failed to listen on a port: %v", err)
-		}
-	}
-	defer ln.Close()
-
+func TestUDPServerCustomDelim(t *testing.T) {
 	conf := NewConfig()
-	conf.TCP.Multipart = true
-	conf.TCP.Address = ln.Addr().String()
+	conf.UDPServer.Address = "127.0.0.1:0"
+	conf.UDPServer.Delim = "@"
 
-	rdr, err := NewTCP(conf, nil, log.Noop(), metrics.Noop())
+	rdr, err := NewUDPServer(conf, nil, log.Noop(), metrics.Noop())
 	if err != nil {
 		t.Fatal(err)
 	}
+	addr := rdr.(*UDPServer).Addr()
 
 	defer func() {
 		rdr.CloseAsync()
@@ -246,94 +223,7 @@ func TestTCPMultipart(t *testing.T) {
 		}
 	}()
 
-	conn, err := ln.Accept()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	wg := sync.WaitGroup{}
-	wg.Add(1)
-	go func() {
-		conn.SetWriteDeadline(time.Now().Add(time.Second * 5))
-		if _, cerr := conn.Write([]byte("foo\n")); cerr != nil {
-			t.Error(cerr)
-		}
-		if _, cerr := conn.Write([]byte("bar\n")); cerr != nil {
-			t.Error(cerr)
-		}
-		if _, cerr := conn.Write([]byte("\n")); cerr != nil {
-			t.Error(cerr)
-		}
-		if _, cerr := conn.Write([]byte("baz\n\n")); cerr != nil {
-			t.Error(cerr)
-		}
-		wg.Done()
-	}()
-
-	readNextMsg := func() (types.Message, error) {
-		var msg types.Message
-		select {
-		case tran := <-rdr.TransactionChan():
-			msg = tran.Payload.DeepCopy()
-			select {
-			case tran.ResponseChan <- response.NewAck():
-			case <-time.After(time.Second):
-				return nil, errors.New("timed out")
-			}
-		case <-time.After(time.Second):
-			return nil, errors.New("timed out")
-		}
-		return msg, nil
-	}
-
-	exp := [][]byte{[]byte("foo"), []byte("bar")}
-	msg, err := readNextMsg()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if act := message.GetAllBytes(msg); !reflect.DeepEqual(exp, act) {
-		t.Errorf("Wrong message contents: %s != %s", act, exp)
-	}
-
-	exp = [][]byte{[]byte("baz")}
-	if msg, err = readNextMsg(); err != nil {
-		t.Fatal(err)
-	}
-	if act := message.GetAllBytes(msg); !reflect.DeepEqual(exp, act) {
-		t.Errorf("Wrong message contents: %s != %s", act, exp)
-	}
-
-	wg.Wait()
-	conn.Close()
-}
-
-func TestTCPMultipartCustomDelim(t *testing.T) {
-	ln, err := net.Listen("tcp", "127.0.0.1:0")
-	if err != nil {
-		if ln, err = net.Listen("tcp6", "[::1]:0"); err != nil {
-			t.Fatalf("failed to listen on a port: %v", err)
-		}
-	}
-	defer ln.Close()
-
-	conf := NewConfig()
-	conf.TCP.Multipart = true
-	conf.TCP.Delim = "@"
-	conf.TCP.Address = ln.Addr().String()
-
-	rdr, err := NewTCP(conf, nil, log.Noop(), metrics.Noop())
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	defer func() {
-		rdr.CloseAsync()
-		if err := rdr.WaitForClose(time.Second); err != nil {
-			t.Error(err)
-		}
-	}()
-
-	conn, err := ln.Accept()
+	conn, err := net.Dial("udp", addr.String())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -358,10 +248,9 @@ func TestTCPMultipartCustomDelim(t *testing.T) {
 	}()
 
 	readNextMsg := func() (types.Message, error) {
-		var msg types.Message
+		var tran types.Transaction
 		select {
-		case tran := <-rdr.TransactionChan():
-			msg = tran.Payload.DeepCopy()
+		case tran = <-rdr.TransactionChan():
 			select {
 			case tran.ResponseChan <- response.NewAck():
 			case <-time.After(time.Second):
@@ -370,12 +259,20 @@ func TestTCPMultipartCustomDelim(t *testing.T) {
 		case <-time.After(time.Second):
 			return nil, errors.New("timed out")
 		}
-		return msg, nil
+		return tran.Payload, nil
 	}
 
-	exp := [][]byte{[]byte("foo"), []byte("bar")}
+	exp := [][]byte{[]byte("foo")}
 	msg, err := readNextMsg()
 	if err != nil {
+		t.Fatal(err)
+	}
+	if act := message.GetAllBytes(msg); !reflect.DeepEqual(exp, act) {
+		t.Errorf("Wrong message contents: %s != %s", act, exp)
+	}
+
+	exp = [][]byte{[]byte("bar")}
+	if msg, err = readNextMsg(); err != nil {
 		t.Fatal(err)
 	}
 	if act := message.GetAllBytes(msg); !reflect.DeepEqual(exp, act) {
@@ -394,23 +291,15 @@ func TestTCPMultipartCustomDelim(t *testing.T) {
 	conn.Close()
 }
 
-func TestTCPMultipartShutdown(t *testing.T) {
-	ln, err := net.Listen("tcp", "127.0.0.1:0")
-	if err != nil {
-		if ln, err = net.Listen("tcp6", "[::1]:0"); err != nil {
-			t.Fatalf("failed to listen on a port: %v", err)
-		}
-	}
-	defer ln.Close()
-
+func TestUDPServerShutdown(t *testing.T) {
 	conf := NewConfig()
-	conf.TCP.Multipart = true
-	conf.TCP.Address = ln.Addr().String()
+	conf.UDPServer.Address = "127.0.0.1:0"
 
-	rdr, err := NewTCP(conf, nil, log.Noop(), metrics.Noop())
+	rdr, err := NewUDPServer(conf, nil, log.Noop(), metrics.Noop())
 	if err != nil {
 		t.Fatal(err)
 	}
+	addr := rdr.(*UDPServer).Addr()
 
 	defer func() {
 		rdr.CloseAsync()
@@ -419,7 +308,7 @@ func TestTCPMultipartShutdown(t *testing.T) {
 		}
 	}()
 
-	conn, err := ln.Accept()
+	conn, err := net.Dial("udp", addr.String())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -445,10 +334,9 @@ func TestTCPMultipartShutdown(t *testing.T) {
 	}()
 
 	readNextMsg := func() (types.Message, error) {
-		var msg types.Message
+		var tran types.Transaction
 		select {
-		case tran := <-rdr.TransactionChan():
-			msg = tran.Payload.DeepCopy()
+		case tran = <-rdr.TransactionChan():
 			select {
 			case tran.ResponseChan <- response.NewAck():
 			case <-time.After(time.Second):
@@ -457,12 +345,20 @@ func TestTCPMultipartShutdown(t *testing.T) {
 		case <-time.After(time.Second):
 			return nil, errors.New("timed out")
 		}
-		return msg, nil
+		return tran.Payload, nil
 	}
 
-	exp := [][]byte{[]byte("foo"), []byte("bar")}
+	exp := [][]byte{[]byte("foo")}
 	msg, err := readNextMsg()
 	if err != nil {
+		t.Fatal(err)
+	}
+	if act := message.GetAllBytes(msg); !reflect.DeepEqual(exp, act) {
+		t.Errorf("Wrong message contents: %s != %s", act, exp)
+	}
+
+	exp = [][]byte{[]byte("bar")}
+	if msg, err = readNextMsg(); err != nil {
 		t.Fatal(err)
 	}
 	if act := message.GetAllBytes(msg); !reflect.DeepEqual(exp, act) {